backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
httpx==0.27.2
python-dotenv==1.0.1
pydantic==2.9.2
python-multipart==0.0.12
trimesh==4.4.9
numpy==1.26.4
torch
open_clip_torch
Pillow
pydicom
nibabel
scikit-image
opencv-python-headless>=4.9.0
mediapipe>=0.10.14
livekit>=1.0.0
livekit-api>=1.0.0
livekit-agents>=1.0.0
pydub>=0.25.1
